Refactor vdr_pi and vdr_control - #114
- Create separate old-style record/play model object, leave plugin
  implementation + main windows life cycle in vdr_pi
- Make control window persistent, use Hide/Show to handle it instead of
  destroy and re-create.
- Create separate control gui interface to break circular dependency
  control window <-> model object
- Clean up model object public interface
- Simplify  toolbar buttons handling.
